+++ LudenWebAPI/Infrastructure/Repositories/GenericRepository.cs
@@ -29,7 +29,7 @@ namespace Infrastructure
                 await context.SaveChangesAsync();
             }
 
-            public async Task RemoveByIdAsync(int id)
+            public async Task RemoveByIdAsync(ulong id)
             {
                 var entity = await GetByIdAsync(id);
                 if (entity != null)
@@ -45,7 +45,7 @@ namespace Infrastructure
                 await context.SaveChangesAsync();
             }
 
-            public async Task<T?> GetByIdAsync(int id)
+            public async Task<T?> GetByIdAsync(ulong id)
             {
                 return await _dbSet.FindAsync(id);
             }
